separate file for ImageControllerCompletionManager
